requirements.txt
azure-cognitiveservices-speech
streamlit
streamlit-mic-recorder
pandas
